+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 ### Added
+- Add intents filter parameter in parsing CLI [#858](https://github.com/snipsco/snips-nlu/pull/858)
 - Add documentation about intents filters [#858](https://github.com/snipsco/snips-nlu/pull/858)
 
 ## [0.20.1] - 2019-09-04